api/pyproject.toml
[tool.poetry]
name = "listany api"
version = "0.1.0"
description = ""
authors = ["Thalida Noel <listany@thalida.com>"]
license = "GNU Affero General Public License"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
django = "^4.1.7"
dj-database-url = "^1.3.0"
graphene-django = "^3.0.0"
pylint-django = "^2.5.3"
psycopg2-binary = "^2.9.5"
django-cors-headers = "^3.14.0"
social-auth-app-django = "^5.2.0"
django-graphql-jwt = "^0.3.4"
gunicorn = "^20.1.0"
django-filter = "^23.1"
whitenoise = {extras = ["brotli"], version = "^6.4.0"}
django-unfold = "^0.5.0"
django-graph-auth = {path = "packages/django-graph-auth-0.1.0.tar.gz"}


[tool.poetry.group.dev.dependencies]
black = "^23.3.0"
isort = "^5.12.0"
pylint = "^2.17.2"
pylint-django = "^2.5.3"
flake8 = "^6.0.0"
pep8-naming = "^0.13.3"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
